Added Collections, Many updates to core code, re organized index.php 's, Cached Config settings, updated plugin handler for better support, added extend library with database hooks. MyTable library is a simplified version of Table, easier to manage, list_view template made for MyTable (see collections).

+++ functions.php
@@ -110,24 +110,26 @@ function get_caller($offset = 0){
 	}
 }
 
-function make_test_dir($dir, $mode = '0755'){
-	if(!is_dir($dir)){
-      mkdir($dir, $mode,true);
-      chmod($dir, $mode);
-   }
-	
-	if(!is_dir($dir)){
-		trigger_error("Do not have write permissions to create directory " . $dir . ". Please change the permissions to allow writing to this directory.");
-		return false;
-	}
-	else{
-		$t_file = $dir . uniqid('test') . '.txt';
-		touch($t_file);
-		if(!is_file($t_file)){
-			trigger_error("The write permissions on $dir are not set. Please change the permissions to allow writing to this directory");
+function _is_writable($dir, $mode = 0755){
+	if(!is_writable($dir)){
+		if(!is_dir($dir)){
+	      mkdir($dir, $mode,true);
+	      chmod($dir, $mode);
+	   }
+		
+		if(!is_dir($dir)){
+			trigger_error("Do not have write permissions to create directory " . $dir . ". Please change the permissions to allow writing to this directory.");
 			return false;
 		}
-		unlink($t_file);
+		else{
+			$t_file = $dir . uniqid('test') . '.txt';
+			touch($t_file);
+			if(!is_file($t_file)){
+				trigger_error("The write permissions on $dir are not set. Please change the permissions to allow writing to this directory");
+				return false;
+			}
+			unlink($t_file);
+		}
 	}
 	
 	return true;
